% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseMSstrings.R
\name{parse_mods}
\alias{parse_mods}
\title{parse_mods
Parse amino acid sequences with embedded protein modifications}
\usage{
parse_mods(seqs, format = "Spectronaut")
}
\arguments{
\item{seqs}{Character vector of amino acid sequences with modification metadata}

\item{format}{Character value specifying the format (only Spectronaut format is currently supported)}
}
\value{
A data.frame with three columns: `sequence`, a list of AAString objects, `mods`, a list of modification metadata, and `mods_at`, a list of locations of the modifications.
}
\description{
parse_mods
Parse amino acid sequences with embedded protein modifications
}
